server: Implement a single workerpool for evolution
The `exportAllToCsvByObject` was the only task that was done in a
workerpool. To support some more, we add a `serverWorkerPool` file to
start the workerpool and a `EvolutionWorkerPool` file that contains the
actual execution code.

The 2 files containing the export to CSV are merged into one and there
is now one function that executes a pool task and another one that
actually does the task in the workerpool.