adminExport: Add functionality to export interview logs to CSV
fixes #721

This adds the functionality to export the interview logs to a CSV file.

There are 2 options for the export:

* `participantResponsesOnly` will export only the data for the fields in
  the `responses` field. Otherwise, all logs will be exported
* `withValues` will export the field as well as the value that has been
  set (or unset)

Both options can be combined.

In the export without values, there is one row per log entry in the
`logs` json field, with `timestamp`, `modifiedFields` and
`unsetFields` columns, along with the interview data (id, uuid,
is_valid, etc), sorted by interview ID, then timestamp.

In the export with values, the is one row per modified value, with a
`timestamp`, `field` and `value` columns, along with the interview data,
sorted by interview ID, then timestamp. Many entries may have the same
timestamp. Unset paths will have empty values.

The export is accessible through a task directly on the server, by
running `yarn node path/to/evolution-backend/lib/tasks/exportInterviewLogs.task.js [--withValues] [--participantResponsesOnly]`.

It is by design not made accessible through the UI, as currently, almost
everyone is admin in the system since we do not have a granular
permission. This contains potentially sensible data that should not be
exposed to everyone.