adminExport: Move the interview log and data files to a sub-directory
As more export files may be available, each set of files should be in a
sub-directory of the `exports` folder to better group them. The
interview raw data files are now in the `interviewData` sub-directory
and the interview logs in the `interviewLogs` directory.

Also, if the requested file does not exist, the server should return a
404 http code instead of 500 as it is not a server error.