frontendHelper: add the `getFormattedDate` and `getGenderedStrings`
`getFormattedDate` is used to format a date in a localized human
readable string. There are options to set the locale, and show the date
with relative day (yesterday, tomorrow, etc) and with/without day of
week

`getGenderedStrings` returns a list of possible strings for the
current person's gender that can be used in the localized strings in the
yml or json files. The gendered string returned are `['suffixE',
'suffixEurRice', 'suffixErEre', 'isHeShe', 'doesHeShe', 'heShe',
'ifHeShe', 'himHerThem', 'hisHerTheir']`. All these gender strings
include a string for the `custom` gender (they/them in english, iel in
french)