survey: scroll to invalid question instead of trying to focus
fixes #740

On Safari on iPhone and iPad, calling the `focus` method on input
elements of type radio or checkbox do not focus on those elements, so if
they are out of view, the participant may not see that there are invalid
widgets.

Since those widget types cannot really "focus" anyway, we simply use the
scrollTo position for all widgets except the text boxes, which work on
all known platforms.