adminExport: differentiate null-initialized from modified data
fixes #738

When getting only the modified fields from the logs, it may be useful to
differentiate between the fields that are initialized to `null` by the
survey from those with actual values, either from the participant
himself or the conditionals. So we add a `initializedFields` column to
the csv file with the names of the fields that are set to `null` in a
log entry.

Also, change the exported file name to include the date, as well as the
parameters with which it was created.