survey: The `startSetInterview` prop type does nto include the `history`
fixes #758

The `startSetInterview` function received in the props receives only 3
parameters that do not include the `History` object. This one is
automatically added by the `dispatch` call. This added the History data
to the survey instead of the prefilled answers, so we have a lot of
un-necessary fields in the interviews.