inputRadioNumber: Make sure the value is always numeric
fixes #765

The inputRadioNumber is for number data. We should not have to specify a
datatype for it to parse the value later, the widget itself should
always return a numeric value or `undefined` when reset.

This also adds test for the handleOnChange call to make sure the
`onValueChange` is called with the expected values.