segments: move the `SegmentsSection` template to evolution-frontend
This template was taken and adapted from the od_nationale_2024 survey.
It displays a summary of each trip with the origin/destination, times
and modes used.

If a trip is selected, its segment widgets are shown as a grouped object,
otherwise each trip has an `edit` button next to it to allow the
selection and a map of a the trips is displayed beside the list.

Add unit tests of the rendering of the section, as well as behavioral
tests for the trip selection.