trips: add a saveTripSegments button widget
This is the button that allows to confirm all segments of a trip have
been entered. It is displayed once the last segment is said to have to
next mode.

Upon saving the trip, it sets the segment's `_isNew` to `false` and
selects the next incomplete trip in the current journey.