survey: add a interviewableAge config and helper function to get persons
The `interviewableAge` configuration option allows to specify the age
from which the persons of the household will be surveyed. Under that
age, the person specific questions will not be asked.

Add a odHelper function `getInterviewablePersonsArray` to retrieve the
persons of the household who have the required age. Unit test this
function.