segments: add widget for person trips title
This widget displays a person trips title, with optionally the nickname
when the person entering the data is not the one being surveyed and the
journey dates.

Add unit tests for the widget

The demo_survey now uses this widget and the journey is initialized with
the tripsDate as startDate.