segments: add the person visited places map widget
This widget displays the visited places and trips on an infoMap, to be
displayed in the UI next to the visited places and segments. The main
code was taken from the od nationale, but it was very similar to the one
in the demo_survey.

There's a lot of FIXMEs with common code and potential
configuration/functionality extracted as we move more widgets in
mainline evolution.

Add unit tests for the widget and replace demo_survey's map widget with
this one.