demo: Add a working demo_generator example
fixes #792

Update the example .xlsx file so it works with the current version of
the generator.

Add the demo_generator example project, empty of all code so it can be
entirely built and compiled at dev time.

Note that the demo_generator is not a complete survey as it is missing
the server bootstrapping files for frontend and backend, it is only
there to make sure the generated code correctly compiles.

To generate the code, run `yarn generateSurvey:example` after having
done the `poetry install` in the `packages/evolution-generator` folder.
Then to compile the generated code and make sure it still works, run the
`yarn workspace demo_generator compile:example` command.