admin: modify `_sections` data before exporting data to csv
fixes #796

For each section, compute/get the total duration, first start timestamp,
number of entries in the section and whether it is completed.

This information replaces the _sections data in the responses to export,
as otherwise it only exports the last value, which does not allow to
calculate real durations.