export interview data: awit for csv stream to end before finishing
Otherwise, the last few records may not be written to the csv file.

Also, check after each write if the buffer is full. If so, pause the
database stream and wait for a drain event on that stream, then resume
the database stream handling. Since there are many streams and each can
be full at a different or same time, we count the number of stream
pauses and wait for the number of reach 0 before resuming the database
stream.