sections: Type the `template` to a string that maps to component
Now that the `SectionConfig` type is in the evolution-common package, it
cannot have a React component as field. Instead, we use a name for the
template and map it to the component in the config.

Add the `templateMapping` application configuration in the frontend that
maps the template names to a template component.

This incidentally allows to provide different section templates for the
admin and participant apps.

The `TripsAndSegmentsSection` template is named
`tripsAndSegmentsWithMap` and the application configures it when it
bootstraps.

The generator has been updated such that the template column can now be
either a boolean or a string. If a boolean, it is a custom template: the
template in the section config will take the value of
`<sectionName>Section` and if the `template.tsx` file does not exist, it
is copied from the generator's typescript skeleton template file. The
section config automatically adds the mapping to the config. If the
template column is a string, it should be one of the builtin template of
evolution and the value will be that string, the mapping should be done
automatically by the platform.

The example has been updated such that the end section has a custom
template. An additional section has been added that uses the builtin
template from evolution 'tripsAndSegmentsWithMap'.

** Breaking change **

Surveys that use section templates need to update their configuration

In the section configuration, replace the `template` field to be a
string instead. For example ` template: SegmentSection` becomes `
template: 'tripsAndSegmentsSection'`.

If the survey provides additional templates, the frontend application's
entry file need to configure the mapping.

For example, in `demo_survey`, the `app-survey.tsx` and the
`app-admin.tsx`  files now contain this code:

```
import appConfig, { EvolutionApplicationConfiguration } from 'evolution-frontend/lib/config/application.config';
import { setApplicationConfiguration } from 'chaire-lib-frontend/lib/config/application.config';

import VisitedPlaceSection from './survey/templates/VisitedPlacesSection';

setApplicationConfiguration<EvolutionApplicationConfiguration>({
    <other application configurations>
    templateMapping: { ...appConfig.templateMapping, 'visitedPlaces': VisitedPlaceSection }
});

runClientApp();
```