objects to csv export: check for the dotted object path
fixes #809

To avoid simple paths being added to an object file because they have
the same prefix (for example `household.personsDidSomething` in the
`household_persons` file, we look for the object path with a '.' (eg.
`household.persons.` in the path.

Also add unit tests for the exportAllToCsvByObject export and fix array
of objects being exported as en empty field with the root object.

Add comments in some places as necessary and remove the `arrayPaths`
field as it is not used or necessary.