segments: Move the segment section configuration in evolution
Task 6 of #501

This moves the segment section configuration as is to evolution-common.

The demo_survey is updated to use this section configuration

Add unit tests for the `preload` and other functions of the section.

The section is not entirely configurable yet, it is just the current
configuration moved to evolution. There are a lot of FIXMEs in the code.
There is still a missing layer between the configurable section and the
actual questionnaire configuration.