conditional: Type the conditional to `ParsingFunction`
The conditional function received in the `checkConditional` function is indeed
a `ParsingFunction`, as defined in the `WidgetConfig` type. This
function does not expect a customPath as third parameter, but a user, so
we fix the call parameters. Also, the calls of the `checkConditional`
effectively send the `user` parameter, which was not the case before.