package.json: add clean and compile script for evolution packages
Often during development, the transition subdirectory is not changed
so does not need to be compiled and cleaned again, unlike the rest of
evolution. So the `yarn clean:evolution` and `yarn compile:evolution`
scripts in the `package.json` file run those commands only in the
evolution workspaces, to avoid needlessly running those commands in
chaire-lib.