widgetOperation: Refactor and document unit tests
The unit tests are complex enough as it is and were using some globally
defined variable, so it was hard to follow what was being done. Now
tests each control the values that they use and more comments are added
above each code block.

Tests are renamed to actually represent what they do, as there was
confusion between validation and conditionals.

Also, each test now calls the `prepareSectionWidgets` function once, but
rather set their previous status instead of relying on a previous call
to set the statuses.