dep: update chaire-lib to latest, with React update
* react: 16.14 => 19.0.0
* react-dom: 16.14 => 19.0
* react-markdown: 8.0.7 => 9.0.1
* react-redux: 7.2.9 => 9.2.0
* react-router-dom: 5.3.4 => react-router: 7.0.2
* react-simple-captcha: 6.1.7 => 9.3.1
* redux: 4.2.1 => 5.0.1
* redux-thunk: 2.4.2 => 3.1.0
* remark-gfm: 3.0.1 => 4.0.0
* @testing-library/react: 10.4.9 => 16.1.0
* history: not used anymore
* react-highcharts: replaced with highchart-react-official 3.2.1

Function and class components that were previously using HOCs were
rewritten to use hooks instead:

* `withTranslation` from `reat-i18next` becomes `useTranslation`
* components that use the redux state now use the `useSelector` hook to
  access a state's value and `useDispatch` to call redux actions.
  instead of the `connect` function

The redux state is better typed.

The `history` package is not required anymore. We use the react-router's
`navigate` function instead. We should eventually rethink the approach
of the redirections from the redux actions, as the `navigate` function
is not suggested in all cases, but only when there is no direct
interaction with the user.

Use the `navigate` and `redirect` functions from `react-router` instead
of using the `history` object from the package of the same name.

Change all components testing from enzyme and react-test-renderer to use
only `@testing-library/react`. This required some changes to how the
components were rendered in unit test, and snapshots needed to be
updated.  Also, many functional tests that were done with enzyme (like
simulating a click or change and checking that the component was
correctly re-rendered) had to be removed, as @testing-library does not
support it as well. We should rather move the playwright for those more
functional tests.

Remove monitoring components that previously used the react-highchart
package, which is not compatible with React 19 anymore and that are
un-used in the demo application, as they cannot be properly tested.

Add the `detectLanguageFromUrl` and `detectLanguage` option
configuration. The first will allow to use `/<lng>` or `?lng=<lng>` in
the URL to automatically change the language. Previously, it used a
regex in the router for this, but that is not supported in the newer
version of react-router.