api: add express middleware to validate uuid format
If the interviewUuid received is of a bad format, the logging middleware
and the query itself will throw exceptions about the format. We add a
proper middleware to validate the interview UUID format before those
calls and return a `400` bad request response to the client instead of
server error.

This middleware is used in validation and user routes that have a uuid
parameter.

Add unit tests for this middleware