survey: refactor the `startSetInterview` redux action
Refactor to use a clearer async function syntax instead of promise. Add
comments on what is being done.

Add unit tests for this action. Change the unit tests to mock the
fetchRetry instead of the node-fetch as it is the method called by the
actions and it's easier to mock a single call. Otherwise, to mock
rejection, we need to mock a specific number of times as the fetchRetry
will retry rejected queries.

Remove the trailing `/` to the server url to activate the interview for
the current user.