survey: decrement the loading state in a finally block
This makes sure all code path decrement the loading state, without
repeating it at every necessary place. All code paths have been tested
and there is no side-effect. The only difference is that if there is no
call to the server to make, the callback is now run before decrementing
the state instead of after, which may have been a mistake in the first
place as the call to the `callback` is done before decrementing the
loading state in the other code path.