styles: move assets and styles to evolution-frontend
And update webpack that should not integrate assets from
evolution-legacy anymore.

** BREAKING CHANGE **

The webpack files, both admin and participant, need to be updated to not
copy the assets directory from evolution-legacy, who does not exist
anymore.

Also make sure all includes of styles and assets in surveys are updated
to use those from evolution-frontend instead of evolution-legacy. Paths
remain the same.