auth: remove references to `connectWithGoogle` and `connectWithFacebook`
These were kept as legacy options from before the config's `auth`
section was added. They are not used anymore and all configs that
previously used it should have been migrated by now.

Also remove the OAuthIntroduction message, which was replaced by a
message at the bottom of the auth page to explain why a login is
necessary in surveys.