admin: let the `SurveyRouter` use the components directly
The conditional loading was correct when there was only one app for
participants and administrative users. Now that the admin app is
separate, we can load the component directly. It will make it easier to
migrate to typescript. And if we need to use conditional loading again
to allow chunk building the app, we can do it later.