Add process_range function to questionnaire_dictionary
This is going to help data analyst to understand the values for input range question.
I did change the choices name to values because it's more simple to understand.
I added a first line for the first line of the csv because it's not a header for all the document.