validation: use typescript helpers in `InterviewStats`
In order to be able to migrate this component, it should use only the
helper functions that are already in the typescript packages.

Note that this component will soon be replaced by one using the audit v2
data, so the validations have simply been replaced with empty errors for
now.