Update generate_labels script
Fix: #625 fix: #518
Rename libelles to labels, because this is the correct term in English.
Use row header name and not number, because the script will be more robust if we need to remove or add an header.
Simplify the generate_labels class to functions. It was hard to understand what was going on.
Half of the stuff was not used or was bugged. So now, we are using fewer functions with better comments to explain everything.
We create deleteYamlFile to fix some bugs when renaming a path (all the paths would stay) and with the order of the questions (there was no order if the file changed)