admin: move the `InterviewStats` component to typescript
Props are typed as they are used, but this component will need to be
migrated to the v2 audit soon, so it was just migrated such that the
demo_survey will work with it, but it is not generic enough for any
survey.

Also add the `svg.d.ts` to allow to import svg images in typescript.