Fix bug with SectionProgressBar
The section completion percentage will now start at 0% when the first section begins and will reach 100% once the last section is initiated.
To achieve this, we will disregard completed sections when calculating the total number of sections.
Test the completed page and the section progress bar in the demo survey.