Setup files that was not generated from Evolution-Generator for demo_generator.
Update the demo_generator README with these limitations.
Make some modifications to the .gitignore to save the git of some of these files.
Update .env.example with the demo_generator, because it will be the main way (soon, only way) of generating a demo.

While most of the files in this directory are generated using the Evolution-Generator, some files were manually created or customized. These include:

-   `config.js`
-   `app-survey.tsx`
-   `server.ts`
-   `helper.ts`
-   `interviewerMonitoring.ts`
-   `roleDefinition.ts`
-   `serverConfiguration.ts`
-   `serverFieldUpdate.ts`
-   `serverValidations.ts`
-   `serverFieldUpdate.test.ts`
-   `serverValidation.test.ts`
-   `common-tests-helper.ts`
-   `test-one-person-no-trips.UI.spec.ts`
-   `test-two-persons-no-trips.UI.spec.ts`
-   `styles.scss`
-   `webpack.admin.config.js`
-   `webpack.config.js`

Use demo_generator inside the package.json, because this will be the main example repo now.
Change the evolution-generator.yml with the good compilation command.