StartUpdateInterview: use named argument instead of positional
The second parameter of the `StartUpdateInterview` function type is not
a `data` parameter with typed arguments. The `valuesByPath`,
`unsetPaths`, interview` and `gotoFunction` are placed in this argument,
to more easily allow for undefined parameters and add eventual new
parameters to this call.

The `callback` function is now the third optional parameter of the call.

Update all usages of this call.

** BREAKING CHANGE **

The `startUpdateInterview` function calls now use a named object
parameter for all function arguments except the callback, who is the
second and last argument.

Example change: previous call: `startUpdateInterview('section', {}, [],
interview, callback, gotoFunction)` now becomes: `startUpdateInterview({
sectionShortname: 'section', valuesByPath: {}, unsetPaths: [],
interview, gotoFunction }, callback)`.