survey update: add field to differentiate the user action from side effects
fixes #814

Add a `userAction` field in the `startUpdateInterview` callback data
parameters.

For now, this user action parameter can either be a `buttonClick` with
the ID of the button clicked by the user, or a `widgetInteraction` with
a widget type and the path and value of the widget clicked. For the
widget interacted with, this path and value replaces what was previously
in the `valuesByPath` object, which now only contains the side-effects
of the user interaction.

The `Button` and `Question` widgets now provide this user interaction
when calling the `startUpdateInterview` callback. And this value is sent
to the backend in the `updateInterview` and the
`updateValidateInterview` calls.

In the backend, the `userAction` parameter is received and we make sure
that a valid request has either a `valuesByPath` or `userAction`
parameter. For now, the interview update simply merges the widget
interaction user action data to the valuesByPath to avoid having to
change this code path right now, as it will be revisited (and
simplified) soon when side effects are in the backend.