Setup the demo_generator for Household_Travel_Survey (#929)
* Setup the demo_generator for Household_Travel_Survey

Update the README of the demo_generator with updated information.
Move generatorConfigs and Household_Travel_Generate_Survey.xlsx to the demo_generator folder, because the references will be closer to the demo.

Don't ignore everything for the demo_generator. We still need the Generator Excel and some tests.

Add prettier inside the demo_generator folder.
Prettier will help keep the code consistent.

Remove an outdated folder inside .gitignore.

Add UI tests for the home section for 1 person and 2 persons household inside demo_generator.
We add the tests before the implementation, to do test-driven development.
Add test:ui command inside package.json

Setup files that were not generated from Evolution-Generator for demo_generator.

Make some modifications to the .gitignore to save the git of some of these files.

Note: We need to compile with 'yarn compile:example'; otherwise, Evolution will compile the evolution-generator when the project is not completely generated.
Note: Demo_Generator needs to be generated before being compiled and built.

Migrate helper inside Evolution-Common and add License
I only keep the helpers that were used inside the demo-generator.
Put the license text inside every TypeScript file
Remove interviwerMonitoring that was not used.

Start from start with serverFieldUpdate and serverFieldUpdate
We just need to add serverFieldUpdate if necessary.
Remove helpers that were not useful.

Synchronize the server.ts of the admin and the client
Rename the file to validationListFilter.ts to make it easier to identify.
Remove serverFieldUpdate.test.ts because there were no tests yet.

Project Config example is at the demo_survey
We want the project config example to be at demo_survey, until the documentation in demo_generator works better.
Have a good import for SetupServer in admin and participant.
Remove google login in example repo, because we don't documented well the process of doing a google login.

While most of the files in this directory are generated using the Evolution-Generator, some files were manually created or customized. These include:

-   `config.js`
-   `app-survey.tsx`
-   `server.ts`
-   `helper.ts`
-   `interviewerMonitoring.ts`
-   `roleDefinition.ts`
-   `serverConfiguration.ts`
-   `serverFieldUpdate.ts`
-   `serverValidations.ts`
-   `serverFieldUpdate.test.ts`
-   `serverValidation.test.ts`
-   `common-tests-helper.ts`
-   `test-one-person-no-trips.UI.spec.ts`
-   `test-two-persons-no-trips.UI.spec.ts`
-   `styles.scss`
-   `webpack.admin.config.js`
-   `webpack.config.js`