Update CI, Landing and Login page for demo_generator
Update CI with playwright tests. For now, we only have the tests for landing page and login page.
Update the survey.yaml for the login and homepage, so we don't have label problems.
Update .gitignore to save the necessary locales files.
Update the package.json and evolution-generator CI with command with the demo-generator workspace, to make is easier to build the demo_generator.
Update the README with updated information.
Update the playwright-example.config with a more robust reporter, it will help find an UI test error.