paradata: add paradata_events table, queries and test
First part of #523

This adds a table that will save one record per paradata event. For
each, it contains:

* interview_id: The ID of the interview touched
* log_timestamp: timestamp at which the log is taken, defaults to now
* user_id: will be the ID of the admin app's user, or the participant if
  the value if `null`
* event_type: The type of this paradata event
* event_data: A json field containing the event data to log

This adds the database query, a simple `log` function for now which will
insert a row for the event, and the tests for this function.

The `logs` field is removed from the `InterviewAttributes` type.

How logs are saved is temporarily updated to add a 'legacy' log at the
same location as previously, to make this commit work without
refactoring the whole paradata logging process yet.

The previous log stream query is temporarily disabled and will come back
in a next commit.