paradata: log paradata events with new log event format
fixes #523

The `updateInterview` function receives a paradata logging function
that can be `undefined` instead of a `boolean` `logDatabaseUpdates`
parameter.

The paradata is logged only once the update is done, and the values
updated by the server are logged independently from the caller's.

Paradata logging is done asynchronously to not block the main process,
but the events are enqueue to ensure sequential processing, as the
insert order is otherwise not guaranteed if there are many parallel
requests.

The loggingMiddleware has one additional function to return the `userId`
for logging purposes, which is expected to be `undefined` when the
request is for the participant.