paradata: update export to add the widget interaction fields
This adds the `widgetType` and `widgetPath` columns to the exported
data. The first will be defined if the event type is
`widget_interaction` and contains the type of widget that was interacted
with. The second will be defined for `widget_interaction` and
`button_click` for respectively the path and button ID interacted with.

When requesting the data with values, the widget interaction user action
is added to the touched fields.