Add home section in demo_generator
Fix tests in home, household, and completed sections.
Add questions inside the home section.
Add dev scripts for the generator to make it easier to develop.
Don't ignore custom files inside demo_generator in .gitignore.
Add some custom files: customConditionals, customGeoData, customValidations, customHelpPopup and customWidgets
Add an incompleted household and completed section for test purposes.
We rename helpPopup.tsx to customHelpPopup to maintain the name convention of manually generated files inside Evolution-Generator.
Rename bikes to bicycles in validations for consistency.
Add GOOGLE_API_KEY inside the CI workflow, to fix the error with Google maps UI tests.
Add a Generator-UI-Test workflow. Update the others workflow. We want the Generator-UI-Test to be able to use the Google maps, but not the other workflows.

BREAKING POINT: Rename customLibelle to customlabel. All the survey with a Google Maps widget and using Evolution-Generator will need to rename the files to be able to see the labels.