sectionConfig: update section config type for repeated blocks
part of #968

Add a `skipIfNoVisibleQuestionWidgets` property to the `SectionConfig`
type to automatically skip a section if no widget of type `Question` is
visible on the page.

Add a `repeatedBlock` property to the `SectionConfig`. The sections that
have this property defined will not be navigatable section, but will
automatically generate a certain number of repeated blocks of section,
depending on the block configuration. For example, some sections could
be repeated for each interviewable household member in an interview. The
repeated objects can be configured.

Update the `parentSection` and `menuName`, which were mutually
exclusive, to show this fact. This forces a section to either be
displayed in the menu or under another section, but not both or neither.

Document each property of the `SectionConfig` type.

Add an empty navigation service with pseudo-code as well as unit tests
for what is the expected behavior.