Fix Python tests for Evolution-Generator
Update the path of all the tests from an old structure version to the new structure.
Fix all the tests, some of the error messages have been changed since the last time we tested.
Ignore some of the generated output and input of these tests.
Install pytest package in Poetry.