ci: add sequential tests to the CI
Using a postgis database, this executes Evolution's sequential test
for every pull request.

Update the participants test so that the IDs of the pre-inserted
participants can use the sequential DB numbering and be saved for later
in the tests. Otherwise, the IDs may clash with the auto-increment ID
from the DB at some point.