response: add the iterationContext property to section navigation
When a section is part of a repeated block, it may be visited many times
with different contexts, depending on the currently selected active
survey object. Instead of saving the complete path of the section, we
separate the section name from the iteration context in the data to
facilitate navigation.