navigation: implement the navigation service
part of #968

It actually write the implementation of the navigation service according
to the spec, manually navigating to the first/last visited section or to
any requested section. It can also navigate forward or backward through
an interview, automatically selecting the active survey object if
required and skipping sections marked to be skipped.

TODO

Handle skipped section to be able to track those sections too