survey: extract a `Survey` component for both admin and participant
`Survey` and `SurveyCorrection` differed only by the values of their
actions and the signature of the startSetInterview redux action.

The common part of both components is thus extracted to a `Survey`
pageParts component and receives the redux action callbacks as props.
The interview is set in the wrapping components and is not required
anymore, as it was only called in the `componentDidMount` method.

Two components `SurveyCorrection` and `SurveyParticipant` set the redux
actions and load the interview respectively for the correction and
participants modes. This results in much smaller components and reduces
code duplication between survey components.

All components are function components.