survey: merge interview create and activation for participants
The `startCreateInterview` and `startSetInterview` redux actions had a
lot of code in common. Instead of duplicating them, we let the server
create the interview in the `activeInterview` route of the participant
app (a 404 is returned in the case of an unexisting interview in the
interviewer app).

Since the function in the common routes does not have a lot to do
anymore, each version of the caller route now does its own code,
instead.

Add unit tests for the participant and user `activeInterview` routes.