survey: add a 'sectionChange' user action
This adds a migration to add the `section_change` type to the event type
in the paradata event table.

Add a type to the `UserAction` type with `sectionChange` as type and a
`targetSection` and `previousSection` fields of type
`NavigationSection`.

Update the code handling the user actions, that was previously doing a
if-else on the 2 available type. Now, functions wrap whatever needs to
be done with those actions (paradata event logging, paradata export,
interview field updates).

When the interview update receives a sectionSection user action, it
"logs" the navigation event by recorded the `_startedAt` for the target
section and adding an action to the `_actions` array. It also set as
completed the previous section if available.

Unit test all the required changes.