survey: use the new navigation service in frontend
fixes #968

Deprecate the `_activeSection` attribute which should not be used
anymore, or set directly (code that supports this use case is still
present in interview update, with an error message indicating to stop
using it).

Add a `navigationService` and `navigation` fields to the survey redux
state to store current navigation history in frontend as well as cache
the navigation service

Add a `startNavigate` redux action which will trigger navigation,
possibly to a requested section (if navigating from a nav link or for
direct request). This action will validate the current section widget
before navigating to the next section. If widgets are valid, it will ask
the navigation service for the next section and update the interview
accordingly.

The `navigationService` is initialized in the `SurveyParticipant` and
`SurveyCorrection` components.

** Breaking change **

While previous code may still work, all code that was previously calling
the `startUpdateInterview` with a `response._activeSection` value to
update should now rather call the `startNavigate` callback. If the
sections are properly configured with their `isSectionCompleted` and
`isSectionVisible` set, navigation should be able to automatically
navigate to the correct next section.

If the survey has repeated blocks (like OD surveys which will ask many
household members to give their trip diaries), the section configuration
should be properly configured with a repeated block containing the
sub-sections. This should reduce the amount of code in the section's
preload code, that was previously used to handle repetition of blocks.