navigation: implement the navigation service
part of #968

Write the implementation of the navigation service according
to the spec, manually navigating to the first/last visited section or to
any requested section. It can also navigate forward or backward through
an interview, automatically selecting the active survey object if
required and skipping sections marked as invisible or disabled.

TODO

Handle skipped section to be able to track those sections too
Update the interview completion percentage calculation
Handle invalid iterations of the repeated block