BREAKING POINT: Rename label in Widgets and Choices sheet.
Rename en and fr label in Widgets and Choices sheet, because this will make it easier to add more languages.
Generate label with {{count}}, {{nickname}} and {{genderedSufix...}} so we can add more logic to the label.
Add all the widgets in household for Demo_Generator.
BREAKING POINT: You need to rename fr to label::fr and en to label::en in Widgets and Choices sheet.