navigation: add `onSectionEntry` and `addSectionExit` in config
These 2 additional functions are executed by the navigation service
respectively on the target section, when navigating to it and on the
previous section, when navigating away from it.

They both return the values to be updated in the interview.

This also deprecates the `preload` function of the `SectionConfig` type.
Everything that was previously done in the `preload` can now be handled
with the `onSectionEntry/onSectionExit` functions when it was value
initialization or with the `isSectionVisible/isSectionCompleted` or
repeated block's `isIterationValid` functions for navigation.