auth: Add an auth method by accessCode/postalCode in the backend
Backend part of #951

Add a project configuration option `auth.authByField`, a boolean value
expressing if this authentication method is supported in this survey.

Add a passport authentication method for a combination of
accessCode and postalCode. It validates if a participant with this field
combination already exists. If so, the user is logged in.

Otherwise, it looks in the prefilled responses table for a matching pair
of accessCode/postalCode and if it exists, it creates and logs the user
with this pair of data.

If the user is not found it returns that the combination is not found.
The participant can then confirm that the values are correct, in which
case, the accessCode/postalCode pair entered will be used as login, even
if there is no matching data.

Also add the express routes to access this authentication method.