Generate labels with other gender
We want to generate labels with 'man', 'woman' and 'other' genders.
Remove getActivePersonId and getActivePersonId helpers functions.
Make a migration for gender values.
Add an adultAge on the config.
BREAKING CHANGE: We don't use 'male' and 'female' gender in survey anymore, because it's is dehumanizing to use 'female' and 'male'
You need to 'yarn migrate' when updating Evolution on a new survey.