survey: add a support form to the participant app
fixes #906

This form is present on every page of the participant app. A button at
the bottom right corner with the text "Need help?" can be clicked by the
participant. It opens a simple form where the user can enter his email
address and a message explaining the support request. The support
request will contain the URL of the current page the participant is on.

Upon submission, this forms sends the message to the backend, which then
sends an email to the addresses specified by the `SUPPORT_REQUEST_EMAIL`
environment variable. This environment variable has the following
format: `lang:comma-separated-emails;lang:comme-separated-emails` to
specify the emails to send the support requests to, as well as the
language that those recipients want the email.

The subject of the support request contains the hostname of the survey
app, to inform the support team which instance is involved. If the
participant is logged in, it will also add the interview ID. The current
page URL, if provided, is also part of the message.

Only the participant app has this form.