survey: pass correct section as props in `GroupObject`
fixes #1030

Passing the group as sectionName was wrong, as the section does not
exist and the widgets were not found when validating the data before the
update, which means that no validation was effectively done.

Also add an error if the requested section does not exist, to catch any
error in section names.