navigation: set the section as completed before navigation
fixes #1032

Since the enableConditional depend on the completion of the current
section, when we navigate forward from a completed section, we need to
set the section as completed, otherwise it won't be possible to navigate
to the next section.

The actual section completion is currently done on the section _after_
the section navigation. We'll need to revisit the section changes, when
implementing #969.